+++ src/__tests__/Bindings_tests.ml
@@ -9,4 +9,13 @@ let () =
           |> Promise.map (function
                | { ChildProcess.exitCode = 0; stdout; stderr = _ } ->
                  expect stdout |> toContainString "hey"
-               | _ -> fail "exec should have succeeded")))
+               | _ -> fail "exec should have succeeded")));
+
+  describe "Expect" (fun () ->
+      let open Expect in
+      testPromise "toBe" (fun () ->
+          ChildProcess.spawn "echo" [| "hey" |] (ChildProcess.Options.make ())
+          |> Promise.map (function
+               | { ChildProcess.exitCode = 0; stdout; stderr = _ } ->
+                 expect stdout |> toContainString "hey"
+               | _ -> fail "spawn should have succeeded")))