+++ src/platform.mli
@@ -19,3 +19,9 @@ module Map : sig
   val find : 'a t -> platform -> 'a
 end
 with type platform := t
+
+type shell =
+  | Sh of Path.t
+  | PowerShell of Path.t
+
+val shell : shell